// theme-generator/theme-colors-generator/src/Interfaces.ts
export interface ButtonState {
  default: {
    background: string
    font: string
    border: string
  }
  hover: {
    background: string
    font: string
    border: string
  }
  pressed: {
    background: string
    font: string
    border: string
  }
  disabled: {
    background: string
    font: string
    border: string
  }
}